requirements.txt
aiogram==3.4.1
redis==5.0.3
python_dotenv==1.0.1
requests==2.31.0